update wallet to apply extrinsics for genesis blocks
